Final draft for now.
Sorry, I had no idea how to pull and branch and push and commit and whatever to the one you had, Ashton. I just copied and pasted yours into a brand spankin new project in VS. This is the Program.cs of that, and is as up-to-date as I had time to do this evening. I added some comments, exception handings, etc.  

I need to go now, but if you guys could look over this, and just keep everyone in the loop as to what changes are needed where, via the old fashioned text message method, that would be great!!